✅ add http_server optional dep test
