+++ pyproject.toml
@@ -40,14 +40,19 @@ caikit-render-interfaces = "caikit.runtime.dump_services:main"
 
 ## Runtime Extra Sets ##
 
+runtime-base = [
+    "prometheus_client>=0.12.0,<1.0",
+]
+
 runtime-grpc = [
+    "caikit[runtime-base]",
     "grpcio-health-checking>=1.35.0,<2.0",
     "grpcio-reflection>=1.35.0,<2.0",
-    "prometheus_client>=0.12.0,<1.0",
     "py-grpc-prometheus>=0.7.0,<0.8",
 ]
 
 runtime-http = [
+    "caikit[runtime-base]",
     "fastapi[all]>=0.100,<1",
     "requests>=2.28.2,<3",
     "sse-starlette>=1.6.1,<2",
