+++ tox.ini
@@ -77,6 +77,26 @@ extras = dev-test
 commands =
     pytest tests/core
 
+# Ensure tests targeting caikit.runtime.grpc_server can be run with no optional dependencies
+[testenv:grpc_server]
+description = run tests against caikit.runtime.grpc_server without anything else
+base = ""
+extras = 
+    dev-test
+    runtime-grpc
+commands =
+    pytest tests/runtime/grpc
+
+# Ensure tests targeting caikit.runtime.http_server can be run with no optional dependencies
+[testenv:http_server]
+description = run tests against caikit.runtime.http_server without anything else
+base = ""
+extras = 
+    dev-test
+    runtime-http
+commands =
+    pytest tests/runtime/http
+
 # Runs dsconverter on caikit
 [testenv:dsconverter]
 description = convert docstrings to google
